DateInputView refined
